Added alignment of legend text and pictograms
Other than MATLAB, Octave allows to change the alignment of
legend text and pictograms using legend('left') and
legend('right'). This commit makes sure that setting is
adhered to in the matlab2tikz output when running Octave.
On MATLAB--which lacks the appropriate setting--legend
entries continue to have the pictograms left of the flush
left text.