Some changes to inline documentation
* Added explanation for 'figurehandle' and 'colormap'
* Corrected explanation for 'filehandle'
* Corrected explanation for 'parseStringsAsMath'
* Some minor changes

It appears that the 'colormap' parameter is not evaluated
and instead the colormap specified by the figure to be
saved is used in all instances. This commit adds a reminder
so that the code may be changed. It also adds another
comment for a possible use of the wrong colormap definition
in the function drawImage.