Apostrophe no longer substituted with \'' in output
Apostrophes used to be substituted along with percentage signs and
backslashes. In the latter two cases this takes place because the
strings are passed through sprintf which would otherwise give them
special meaning. In the case of apostrophes this is not the case.
Only when writing out apostrophes they have to escaped like so:

  string = 'before''after'

For more information on string escaping see
http://www.mathworks.com/help/techdoc/ref/fprintf.html#inputarg_formatSpec