Added test function `texrandom'
This test case uses most symbols known to MATLAB and generates five
random strings from them. Grouping braces and sub/superscript
operators are then inserted into the strings in such a way that it
is a valid MATLAB TeX string.

Due to the random nature of the test, its successful completion
does not indicate absence of matlab2tikz bugs, but in running it
multiple times it can also produce combinations that wouldn't all
be generated by hand. Its "inhuman" nature makes it particularly
likely to point out bugs that would not show up with more natural,
human-generated strings.

The test prints out the strings it generated to the console. Since
these change with every iteration of the test, it's imperative to
write them down in case of a test failure.