extraAxisOptions can overwrite matlab2tikz's axis options
When setting for instance "axis on top=false" via extraAxisOptions,
this used to be overwritten by matlab2tikz setting this option to
true after the manual user setting. Changing the order in which
options are appended makes the user-provided extraAxisOptions gain
priority over the options generated by matlab2tikz.

This fixes github issue #77