Fix \figureheight regression
Commit e3cb8f88 introduced a different way to extend cell arrays.
This caused an error at least in Octave 3.6.1 if matlab2tikz was
called with the parameter height='\figureheight' (or something
else starting with a backslash) because in that case the initial
value assignment was to a string, not a cell. When the cell
element 'scale only axis' was supposed to be added afterwards,
Octave aborted with an error, saying "sq_string cannot be indexed
with {".

This commit fixes that bug.