Fix basic legend handling in Octave.
This fix changes the legend handling so instead getting an error, there is
basic support for legends in Octave. The support is probably not complete,
but all basic tests in Octave result in getting the legend texts exported.

This commit should not change Matlab's behaviour, but the fix has not
been tested with Matlab.