VersionControlIdentifier
When possible, the commit identifier is added to the output.
The implemented method relies upon either:
 - a .git directory present in the directory above matlab2tikz.m
   (as is the case in Nico's repository)
   this should contain a HEAD file (standard for a git repo)
 - a .commitinfo file present in the directory containing matlab2tikz.m
   (when the above is not present)

The tree-ish in this file is read (and followed when it is of the form ref:...
 until it is no longer a dynamic reference OR an indirection limit is reached).

The method does NOT rely on git being installed on the user's system, but relies
directly on the simple file structure of a git repository.

All errors are caught and result in no commit information to keep this
auxilary information from breaking matlab2tikz as a whole.