Add support for different length up and down error bars
This patch utilizes new features in pgfplots 1.9 that allow
different length of up and down errorbars.  The coordinate
input format changes from:
  (1,21.35) +-= (0.0,8.65)
to (for example):
  (1,21.35) += (0.0,8.65) -= (0.0,10.32)

To be honest I'm not exactly sure if this patch has any side-
effects, due to lack of time it hasn't been properly tested,
but works for my application, which is a matlab plot generated
by the command:
  errorbar(X, Y, ED, EU)
where X, Y, ED and EU are simple vectors.