Default handling: use back-tracking
When an unknown handle graphics element is encountered,
execute the default handling using back-tracking.
Whenever an error occurs, roll back the state of `m2t` to before
encountering the hg-object and continue with the next element.

The original error message is shown for debugging purposes, but
this allows one to run `matlab2tikz` on figures that have
unsupported elements, as was the case in Issue273