Only check for updates when not using git
Checking for updates when using git is a hinderance, e.g.
when trying to run a git bisect as then it breaks up the run
when jumping accross versions.

Not checking for git users seems like a good alternative
as we can assume those users already check for new versions using git