Color mixtures implemented
Basic idea behind it is to run over all defined colors Ci and Cj
and find a value of `p` that reconstructs rgb as Ck according
to the color mixing equation of xcolor.sty

I.e. in xcolor `Ci!P!Cj` with P = 100*p means

    Ck = p*Ci + (1-p) * Cj

or, equivalently:

   Ck - Cj = p*Ci - p*Cj = (Ci - Cj) * p

This can be solved for `p` in MATLAB using:

   p  = (Ci-Cj) \ (rgb(:) - Cj);

where `rgb` is approximated by `Ck` in these equations.
We round the value of `p` to a percentage, as in xcolor.sty and see
how well the approximation of `rgb` by `Ck` holds in infinity norm sense.
This should yield a more conservative guess than the two-norm/least-squares
solution that is computed by the backslash operator.
But the infinity norm `max(abs(Ck-rgb))` is easier to compute and comprehend
than the actual two-norm `sqrt(mean((Ck-rgb).^2))`.
Practically, it means that for some colors `rgb`, there might be a close
approximation `Ck` that is rejected due to one of its RGB components being
off by a lot. So when in doubt, a color is defined instead of mixing two others.