Externalize all tables
All tables are put in separate external TSV files (tab separated values).
This can be toggled by using the `externalData` flag which is default on.

As with PNG files, they can be specified using a relative directory.
A new switch `relativeDataPath` is therefore introduced to take the
role of `relativePngPath`, which should be deprecated in a future release

Something to keep in mind:

 * the row separator is different for interal tables (`\\\\\n`, i.e. a
   LaTeX newline and newline character) or external tables (`\n`,
   only a newline character)

Also, perhaps it may be suited better to let `makeTable` add options
to the `table[...]` command, e.g. detection of headers, separators, ...
to keep code duplication to a minimum