Color depth reduced to 48 bit
60 bit was overkill anyhow. 48 bit is a more reasonable setting
as this is sometimes used in professional software (e.g. Adobe (R)).

Also, it reduces the number of digits from 7 to 5, such that
storage is reduced (and readability of the TikZ code is increased).
This also makes it somewhat easier to make color mixtures, as the precision
is reduced tenfold.