test: less MATLAB PDF clutter
MATLAB part of the ACID test reference contained a lot of clutter,
that may ineffectively scale a PDF file as to obtain a properly
cropped PDF file.

Truth is, at the moment it is probably not possible to write robust and
compact code to make MATLAB do this. Especially if subplot figures are
used, this becomes nearly impossible. It is better to keep EPS
with its downsides (e.g. no alpha channel support, some background colors
may be lost, ...) and convert the file later on