Make: don't do epstopdf
EPSTOPDF should be called by LaTeX, not by the Makefile.

This is done by the `epstopdf` package.
Therefore, when a proper LaTeX installation exists, the test will run.
We get rid of nastiness of setting the PATH and making sure that everything
is accessible from the PATH.

With the current approach, when LaTeX is accessible from the PATH, calling
epstopdf becomes a responsibility of the LaTeX distribution. This means
that we shouldn't care about the PATH: either LaTeX can be called and is
properly configured and everythng will be fine. If LaTeX can't be called,
running epstopdf is useless anyhow, and is unlikely to succeed anyway.