Quiver R2014b added
In R2014b, the line series with explicit coordinates is no longer present.
This means that it was necessary to re-implement the scaling algorithm of MATLAB
Consequently, the previous shortcut has been removed, making sure that the code
base is equivalent for the different versions of MATLAB.

The output from the ACID tests (23 and 25) looks fine from R2014b. Older
versions have not yet been tested, that will happen right before merging this
branch into master