Patch plots should be cycled
Some plots suffer from not having a `--cycle` at the end, causing
the first and last point not to be connected, e.g. #317.
In the code it was suggested that some do not need this, but for
none of our test cases, this change produces errors.

Some data sets, however, do contain the same first and last point.
In those cases, adding a cycle is superfluous, but doesn't hurt :-)