Single output in `testfunctions`
The testfunctions were riddled by a lot of code to set
optional output arguments to be passed to `m2t` and `cleanfigure`.
This is circumvented by passing a single struct as output.
Once the struct is obtained, values that are not set are filled
to their default values (empty).

This allows to remove quite some code:

 * in each test case most output arguments and `extraOpts = {}`.
 * in `testfunctions` the hacky approach to determine the number
   of outputs. Now it's allways 1.

Moreover, clarity has improved since now we can pass `stat.skip = true` to
signal that a test should be skipped or `stat.closeall` to call `closeall`.
Before, this was signaled using resp. `description = ''` and using an outside
array.