Use `opts_append` and fix `opts_merge`
`opts_append` needs to be used where the colormap is handled.
So, in `opts_merge` this also has to be done.
Since previously, `opts_append` did no checking at all, now it is checked
whether the (k,v) pair exists already. If it does, there is no point in adding
it again.
Note: two pairs (k,v1) and (k,v2) WILL be added.
As this needed the introduction of functions `opts_has` and `opts_get`,
the same are used in `opts_add` which has now been simplified