Test Harness: fix `extraFunctionFromError`
When the error occurs in the upper stack frame, handling
was impaired, because the upper+1 stack frame was queried.
Instead, now we first initialize the function name to '' and
explicitly check before indexing.

In that case, the function name is not returned, but it is also very
likely the function has not even been called.