Acceptable complexity flag for `codeReport`
To ignore the complexity of a particular function, place `#COMPLEX`
in its help text. It is advisable to also state a reason why this should be
acceptable, e.g. `#COMPLEX: big switch-case`.

By putting the flag, the complexity is ignored in the statistics computed
in `codeReport` and it is plotted distinctly in that report.