Clean ACID test output before running
This can be toggled using the `cleanBefore` flag.
Internally, `make clean` is used to do so.
This has the downside that on any system without a make
on the path (most installs of Windows), this does not work.
However, this downside is mitigated by the `try-catch`, such that
on improperly set-up systems, this doesn't break execution of the tests.