Fix `cleanfigure`'s `handle` validator
USing `ishandle` to determine whether a handle is being passed.
Before, `isnumeric` was used, this is wrong and breaks in R2014b.

The right function to use would be `ishghandle`, but since I'm unsure this works
in Octave or old MATLAB versions, `ishandle` is used.
The difference between both should only be noticable for Simulink objects,
which is outside our scope anyhow