Use `assertRegularAxes`
The replaced `switch-case` was nonfunctional in:
 - Octave 3.8.0
 - MATLAB R2014a
 - MATLAB R2014b
for our ACID testsuite: only the last branch was ever entered.
So it can be removed.

However, for safety, I've introduced the function `assertRegularAxes`
that makes sure that only that branch would ever be called.

In the odd case that some code would need this switch-case, we should
get an issue report in the future.