MD5 in MATLAB
Using a Java-based implementation from
  http://stackoverflow.com/questions/12140458/calculating-md5-hash-rfc-1321-conform-in-matlab-via-java

While technically, its license is CC-BY-SA 3.0 (by StackOverflow policy),
the snippet is small and generic enough to count as an excerpt for which
the license can be ignored.
This is also the the interpretation Jeff Atwood, so the actual point of view of
StackOverflow on the matter.

I attributed the code to the respective people, put a link to SO, so that should
be covered. This code has been moved into an external file, should there be any
problem with the licensing, that it can be easily removed.