Remove `hasLines` from `splitByArraySize`
This was needed to determine whether or not the last point of a data chunk is
repeated to stitch data chunks together when there is a line connecting
the points.
Instead, just put the lonely point at all times: the overhead is so
small that it doesn't justify the added complexity in the code.