Replace `passReliable` -> `passUnreliable`.
Unreliable tests that pass are actually more interesting.
E.g. in the case that we try to make more test cases reliable,
this will be the variable that we are going to compare amongst
environments.

`passReliable` can be constructed easily using
 `setdiff(reliable, failReliable)` and similarly for
 `failReliable`. As such, they are not needed.

I could say the same about `reliable` and `unreliable`,
but there it might be of interest to see these values directly
from the output, e.g. on Travis.