Replace findobj + set(0, 'ShowHiddenHandles')
`set(0, 'ShowHiddenhandles', 'on')` sets global state.
Instead, the same effect can be gotten with using:
 - `findall` instead of `findobj`
 - `allchild(h)` instead of `get(h, 'Children')`

In the end, maybe we can replace some `findall` commands with
`findobj` again, if we know that hidden handles do not occur in that case.
However, that's just a finishing touch and not a fundamental problem.