Allow patchcolor 'none'
Don't throw an error here, to fix #739.
We don't have any test cases that cover this branch anyway.

I really wonder whether that error is really needed. I ran a `git blame`
on this and I tracked along the following commits:

 * 1a0e89ee where I just change some whitespace formatting
 * 20d9fc8e (whitespace above the error)
 * 9fbf9363 where Nico already has a similar error message in place
   but the function is used for a generic color conversion (instead of patches).

So my presumption is that the code was changed, the error message carried over
without a full check on whether it was really needed.

Anyway, the current change MAY cause compile errors if I'm being to quick in
removing the error. I expect we will see some bug reports if that's the case,
so I left a comment in the code to help tracking down this information.