`opts_print(m2t, ...)` -> `opts_print(...)`
There was no need for the `m2t` parameter, since
all entries get formatted as a string beforehand anyway.

As such, we can fall back on the default string format in `m2tstrjoin`
which is currently `%g` as it is not used.

This change cleans up the syntax a bit