Add support for categorical values in bar plots
Categorical values can be plotted in tikz by defining the values as
symbolic, and passing the data as a set of coordinates. This commit
implements a function to format data as coordinates, and it adds some
checks to handle categorical values in bar plots (previously, trying to
plot a bar plot with categorical values generated an error).