With this commit I address: ISSUE 1032
%%%  Argument check functions %%%

I added three validator functions that are simmilar to previousely used
validator functions. But now they also accept also matlab Strings.
One can see the new functions in the following table together with
the functions that were used previousely:

| previousely     | now also accepting matlab strings |
|---		  |---				      |
| IsChar          | isCharOrString		      |
| isCellOrChar    | isCellOrCharOrString	      |
| isCellstrOrChar | isCellstrOrCharOrString	      |

The script in which I checked their functionalities are:
- src/exampleArguments2Char.m

%%% Arguments conversion fuctions %%%

Additionally I added the function convertArgumetns2Char which converts
all input variables  that are string to char. If the inputs are not
strings the variables are left unchanged.

The script which tests the functionalities of convertArguments2Char is:
- src/exampleArguments2Char.m

%%% Application in matlab2tikz.m and m2tcustom.m
I added the before explained functions to matlab2tikz.m as well as m2tcustom.m.

I tested the functionallity in the function using a slightly changed test/examples/example_bar_plot.m.
I admit this test is not very good. Inputs for a better test would be appreciated

short summary:
	new file:   src/exampleArguments2Char.m
	new file:   src/exampleIsFunctions.m
	modified:   src/m2tcustom.m
	modified:   src/matlab2tikz.m
	modified:   test/examples/example_bar_plot.m